backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy[asyncio]==2.0.30
asyncpg==0.29.0
alembic==1.13.1
pydantic==2.7.1
pydantic-settings==2.2.1
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.9
anthropic==0.26.0
openai==1.30.1
PyPDF2==3.0.1
python-docx==1.1.2
redis==5.0.4
httpx==0.27.0
aiofiles==23.2.1
tiktoken==0.7.0
google-generativeai==0.8.3
beautifulsoup4>=4.12.0
lxml>=5.0.0
trafilatura>=2.0.0
pyotp>=2.9.0
qrcode>=7.4.0
stripe>=9.0.0
pytest>=8.0.0
pytest-asyncio>=0.23.0
aiosqlite>=0.20.0
openai-whisper>=20231117
